Release: Canvases and Notifications

+++ src/domain/common/canvas/canvas.service.ts
@@ -1,23 +1,52 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { EntityNotFoundException } from '@common/exceptions';
+import { FindOneOptions, Repository } from 'typeorm';
+import {
+  EntityNotFoundException,
+  NotSupportedException,
+} from '@common/exceptions';
 import { LogContext } from '@common/enums';
 import { Canvas } from './canvas.entity';
 import { ICanvas } from './canvas.interface';
-import { UpdateCanvasInput } from './canvas.dto.update';
+import { CreateCanvasInput } from './dto/canvas.dto.create';
+import { AuthorizationPolicy } from '../authorization-policy/authorization.policy.entity';
+import { CanvasCheckoutService } from '../canvas-checkout/canvas.checkout.service';
+import { UpdateCanvasInput } from './dto/canvas.dto.update';
+import { CanvasCheckoutLifecycleOptionsProvider } from '../canvas-checkout/canvas.checkout.lifecycle.options.provider';
+import { ICanvasCheckout } from '../canvas-checkout/canvas.checkout.interface';
 
 @Injectable()
 export class CanvasService {
   constructor(
     @InjectRepository(Canvas)
-    private canvasRepository: Repository<Canvas>
+    private canvasRepository: Repository<Canvas>,
+    private canvasCheckoutService: CanvasCheckoutService,
+    private canvasCheckoutLifecycleOptionsProvider: CanvasCheckoutLifecycleOptionsProvider
   ) {}
 
-  async getCanvasOrFail(canvasID: string): Promise<ICanvas> {
-    const canvas = await this.canvasRepository.findOne({
-      id: canvasID,
+  async createCanvas(canvasData: CreateCanvasInput): Promise<ICanvas> {
+    const canvas: ICanvas = Canvas.create(canvasData);
+    canvas.authorization = new AuthorizationPolicy();
+
+    // get the id assigned
+    const savedCanvas = await this.save(canvas);
+
+    canvas.checkout = await this.canvasCheckoutService.createCanvasCheckout({
+      canvasID: savedCanvas.id,
     });
+    return await this.save(canvas);
+  }
+
+  async getCanvasOrFail(
+    canvasID: string,
+    options?: FindOneOptions<Canvas>
+  ): Promise<ICanvas> {
+    const canvas = await this.canvasRepository.findOne(
+      {
+        id: canvasID,
+      },
+      options
+    );
     if (!canvas)
       throw new EntityNotFoundException(
         `Not able to locate Canvas with the specified ID: ${canvasID}`,
@@ -27,11 +56,33 @@ export class CanvasService {
   }
 
   async deleteCanvas(canvasID: string): Promise<ICanvas> {
-    const canvas = await this.getCanvasOrFail(canvasID);
-    return await this.canvasRepository.remove(canvas as Canvas);
+    const canvas = await this.getCanvasOrFail(canvasID, {
+      relations: ['context', 'checkout'],
+    });
+    // check it is a canvas direction on a Context
+    if (!(canvas as Canvas).context) {
+      throw new NotSupportedException(
+        `Not able to delete a Canvas that is not contained by Context: ${canvasID}`,
+        LogContext.CHALLENGES
+      );
+    }
+    if (canvas.checkout) {
+      await this.canvasCheckoutService.delete(canvas.checkout.id);
+    }
+    const deletedCanvas = await this.canvasRepository.remove(canvas as Canvas);
+    deletedCanvas.id = canvasID;
+    return deletedCanvas;
   }
 
-  updateCanvas(
+  async updateCanvas(
+    canvas: ICanvas | undefined,
+    updateCanvasData: UpdateCanvasInput
+  ): Promise<ICanvas> {
+    const updatedCanvas = this.updateCanvasEntity(canvas, updateCanvasData);
+    return await this.save(updatedCanvas);
+  }
+
+  updateCanvasEntity(
     canvas: ICanvas | undefined,
     updateCanvasData: UpdateCanvasInput
   ): ICanvas {
@@ -42,6 +93,39 @@ export class CanvasService {
       );
     if (updateCanvasData.name) canvas.name = updateCanvasData.name;
     if (updateCanvasData.value) canvas.value = updateCanvasData.value;
+    if (updateCanvasData.isTemplate !== undefined)
+      canvas.isTemplate = updateCanvasData.isTemplate;
     return canvas;
+  }
+
+  async save(canvas: ICanvas): Promise<ICanvas> {
+    return await this.canvasRepository.save(canvas);
+  }
+
+  async getCanvasCheckout(canvas: ICanvas): Promise<ICanvasCheckout> {
+    const canvasWithCheckout = await this.getCanvasOrFail(canvas.id, {
+      relations: ['checkout'],
+    });
+
+    // Note: this is done the same as with organizationVerification, i.e. that
+    // we create the entity wrapping a lifecycle if it is not present.
+    // This is due to lifecycles being difficult to deal with via migraitons.
+    // Needs further discussion.
+    if (!canvasWithCheckout.checkout) {
+      // create and add the checkout
+      canvas.checkout = await this.canvasCheckoutService.createCanvasCheckout({
+        canvasID: canvas.id,
+      });
+
+      await this.save(canvas);
+    }
+
+    if (!canvasWithCheckout.checkout)
+      throw new EntityNotFoundException(
+        `Canvas not initialised, no checkout: ${canvas.id}`,
+        LogContext.CONTEXT
+      );
+
+    return canvasWithCheckout.checkout;
   }
 }
